backend_webapp/requirements.txt
Django==4.0
djangorestframework==3.12.4
boto3==1.21.5
